move file-put tests to its own module.
Rationale: file-get tests should refer to an immutable data set. Therefore, we need to point our bucket setup at a different location. Since the client is set up in `setUp`, we need to have a different `setUp` method for read and write tests.
