+++ tests/test.py
@@ -5,7 +5,6 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import os
 import sys
-import uuid
 import unittest
 
 import requests
@@ -35,26 +34,6 @@ class TestDSS(unittest.TestCase, DSSAsserts):
         self.assertGetResponse(
             "/v1/files/91839244-66ab-408f-9be5-c82def201f26?replica=aws",
             requests.codes.found)
-
-    def test_file_put(self):
-        file_uuid = uuid.uuid4()
-        response = self.assertPutResponse(
-            "/v1/files/" + str(file_uuid),
-            requests.codes.created,
-            json_request_body=dict(
-                source_url="s3://hca-dss-test-src/test_good_source_data",
-                bundle_uuid=str(uuid.uuid4()),
-                creator_uid=4321,
-                content_type="text/html",
-            ),
-        )
-        self.assertHeaders(
-            response[0],
-            {
-                'content-type': "application/json",
-            }
-        )
-        self.assertIn('timestamp', response[2])
 
     def test_bundle_api(self):
         self.assertGetResponse("/v1/bundles", requests.codes.ok)
